requirements.txt
pandas
numpy
scikit-learn
pyyaml
plotly
streamlit
pytest
yfinance
pyarrow
exchange-calendars
